feat(mouth): restore balizero.com home to team + 4 funnel design
The 2026-04-17 commit ea33987d9 swapped the marketing homepage to a
static editorial carousel of 5 hardcoded articles dated Apr 2, 2026.
That carousel never auto-updated as new articles shipped, so by
2026-04-27 the homepage looked frozen 25 days behind production.

Restore the team-photo hero + four funnel feature cards (visa, kbli,
tax, property) + SocialProof + dynamic NewsHero/LatestNews — same
design that has been live under /v2 since April. The /v2 page is
preserved as-is for design-system reference.

The new (marketing)/page.tsx imports from ../v2/_components/*
(MobileNav, HeroBlueprint, SocialProof, FunnelFeature, NewsHero,
TopicPills, LatestNews, Footer, ZantaraFAB) and reuses the existing
homepage-layout.json mapping for hero/latest article slugs.

Metadata kept from the previous home: canonical balizero.com, no
noindex (the /v2 variant has noindex; the production root must
remain indexable). The editorial carousel is preserved as
page.editorial-backup.tsx for rollback if needed.

Verified: tsc --noEmit clean, next build OK.
